- missed that message.info.response.http_version should also be double

+++ php_http_message_api.h
@@ -41,7 +41,7 @@ struct _http_message {
 		} request;
 
 		struct {
-			float http_version;
+			double http_version;
 			int code;
 		} response;
